+++ src/lib/ui/AccessCodeForm/AccessCodeFormTimezonePicker.tsx
@@ -1,14 +1,7 @@
 import { useState } from 'react'
 
-import {
-  getBrowserTimezone,
-  getTimezoneLabel,
-  getTimezoneOffset,
-  getTimezones,
-} from 'lib/dates.js'
-import { Checkbox } from 'lib/ui/Checkbox.js'
 import { ContentHeader } from 'lib/ui/layout/ContentHeader.js'
-import { handleString } from 'lib/ui/TextField/TextField.js'
+import { TimezonePicker } from 'lib/ui/TimezonePicker/TimezonePicker.js'
 
 interface AccessCodeFormTimezonePickerProps {
   value: string
@@ -21,44 +14,19 @@ export function AccessCodeFormTimezonePicker({
   value,
   onClose,
 }: AccessCodeFormTimezonePickerProps): JSX.Element {
-  const [manualTimezoneEnabled, setManualTimezoneEnabled] = useState(false)
-
-  const isBrowserTimezone = value === getBrowserTimezone()
-  const isManualTimezone = !isBrowserTimezone || manualTimezoneEnabled
-
-  const handleChangeManualTimezone = (enabled: boolean): void => {
-    setManualTimezoneEnabled(enabled)
-    if (!enabled) {
-      onChange(getBrowserTimezone())
-    }
-  }
-
-  const title = isManualTimezone ? t.titleManual : t.titleAuto
+  const [title, setTitle] = useState(t.titleAuto)
 
   return (
-    <div className='seam-timezone-picker'>
+    <div className='seam-access-code-timezone-picker'>
       <ContentHeader title={title} onBack={onClose} />
       <div className='seam-content'>
-        <Checkbox
-          label={t.setTimezoneManuallyLabel}
-          checked={!isManualTimezone}
-          onChange={(manual) => {
-            handleChangeManualTimezone(!manual)
-          }}
-          className='seam-manual-timezone-checkbox'
-        />
-
-        <select
+        <TimezonePicker
           value={value}
-          onChange={handleString(onChange)}
-          className='seam-timezone-select'
-        >
-          {getTimezones().map((timezone) => (
-            <option value={timezone} key={timezone}>
-              {t.utc} {getTimezoneOffset(timezone)} {getTimezoneLabel(timezone)}
-            </option>
-          ))}
-        </select>
+          onChange={onChange}
+          onIsManualTimezoneChanged={(isManualTimezone) =>
+            setTitle(isManualTimezone ? t.titleManual : t.titleAuto)
+          }
+        />
       </div>
     </div>
   )
