+++ src/styles/_access-code-form.scss
@@ -117,23 +117,9 @@
 }
 
 @mixin timezone-picker {
-  .seam-timezone-picker {
+  .seam-access-code-timezone-picker {
     .seam-content {
       padding: 0 24px;
-
-      .seam-manual-timezone-checkbox {
-        margin-bottom: 8px;
-      }
-
-      > .seam-timezone-select {
-        background: colors.$white;
-        padding: 0 12px;
-        width: 100%;
-        border: 1px solid colors.$text-gray-3;
-        border-radius: 8px;
-        height: 48px;
-        font-size: 16px;
-      }
     }
   }
 }
