+++ src/styles/_timezone-picker.scss
@@ -1,0 +1,22 @@
+@use './colors';
+
+@mixin all {
+    .seam-timezone-picker {
+  
+  
+        .seam-manual-timezone-checkbox {
+          margin-bottom: 8px;
+        }
+  
+        > .seam-timezone-select {
+          background: colors.$white;
+          padding: 0 12px;
+          width: 100%;
+          border: 1px solid colors.$text-gray-3;
+          border-radius: 8px;
+          height: 48px;
+          font-size: 16px;
+        }
+      }
+    
+  }
